Fixing issues with editing stars & Fixing Issue #11
https://github.com/chaosdrop/FlatNotes/issues/11

Turns out this was actually old debut code that still needed to be used to fix an issue with not being able to change stars after doing a bulk star edit.

The issue was that since the GUI used for the star edit box is used in multiple places it was getting locked into an exit loop.