Fixing issue #13
https://github.com/chaosdrop/FlatNotes/issues/13

Added a check and then return to prevent blank ghost notes from being created by accident.

I also tried making it so a new note is made from the search term, but that fails because of the delayed nature of the saving. I left the code there commented out for now.