Remove everything related to treeview
Properly remove all the treeview code since it was too much hassle to deal with in AHK. This commit should have all of the code for the treeview if I ever want to try again, but I doubt it. I'd rather spend time rebuilding flatnotes in a way that supports OSX than spend time on a tree feature.