Added option to fix note details
This fixes manually deleted bookmarks, images, and clips as those details are kept in the ini file and not updated when a person manually deletes them.

I may change this to be done on load at some point but it seems kind of expensive for large note collections to do every load.