Option to Sync Changes
In theory, this is useful by itself, but it's mainly going to be used for an auto-sync system later. The idea is that Flatnotes will check a file that will have the name of the last computer that saved changes. If that name is different than the host computer name it'll auto refresh.