Automatic Network Syncing.
FlateNotes now checks the _sync.ini to see if any other computer have saved a file and will refresh the note lists. This makes it as easy as hitting the hotkey to show the library to update a network sync.

It may have issues in slow locations, but we'll see. I wouldn't recommend using a flat note program on super slow storage or networks anyway.